otus_project/requirements.txt
kafka-python
flask
prometheus_client
clickhouse-connect
gunicorn